requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
dj-database-url==0.5.0
Django==3.2.9
django-heroku==0.3.1
gunicorn==20.1.0
importlib-metadata==3.10.1
psycopg2==2.9.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
whitenoise==5.3.0
zipp==3.6.0
